finish separating page tests
